fixed-point: converted QMF functions
git-svn-id: http://svn.xiph.org/trunk/speex@5438 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/filters.h
@@ -47,7 +47,7 @@ typedef struct CombFilterMem {
 } CombFilterMem;
 
 
-void qmf_decomp(float *xx, float *aa, spx_sig_t *y1, spx_sig_t *y2, int N, int M, float *mem, char *stack);
+void qmf_decomp(short *xx, float *aa, spx_sig_t *y1, spx_sig_t *y2, int N, int M, float *mem, char *stack);
 void fir_mem_up(spx_sig_t *x, float *a, spx_sig_t *y, int N, int M, float *mem, char *stack);
